Add feature flag for componentized map
