+++ app/src/UI/LegacyMap/Controls/AccuracyToggle.tsx
