+++ app/src/UI/LegacyMap/Controls/CenterCurrentRecord.tsx
