+++ app/src/UI/LegacyMap/Controls/CustomizeLayerDialog.tsx
