+++ app/src/UI/LegacyMap/Controls/MapToggleCacheGeoJSON.tsx
