+++ app/src/UI/LegacyMap/Controls/NewRecord.tsx
