+++ app/src/UI/LegacyMap/Controls/PauseGeoTrackingButton.tsx
