+++ app/src/UI/LegacyMap/Controls/QuickPanToRecordToggle.tsx
