+++ app/src/UI/LegacyMap/helpers/draw-tools.ts
