+++ app/src/UI/LegacyMap/helpers/utility-functions.ts
