+++ app/vite.config.ts
@@ -31,6 +31,8 @@ function buildSpecificDefines() {
     defines['CONFIGURATION_PUBLIC_MAP_URL'] = JSON.stringify(process.env['PUBLIC_MAP_URL']);
     defines['CONFIGURATION_IAPP_GEOJSON_URL'] = JSON.stringify(process.env['IAPP_GEOJSON_URL']);
     defines['CONFIGURATION_SILENT_CHECK_URI'] = JSON.stringify(process.env['SILENT_CHECK_URI']);
+
+    defines['CONFIGURATION_COMPONENTIZED_MAP'] = JSON.stringify(process.env['ENABLE_COMPONENTIZED_MAP']);
   } else if (process.env.CONFIGURATION_SOURCE === 'Caddy') {
     defines['minify'] = false;
